+++ tiger-testenv-mgr/src/frontend/src/components/global_configuration/EditActionButtons.vue
@@ -1,0 +1,12 @@
+<!--
+  - ${GEMATIK_COPYRIGHT_STATEMENT}
+  -->
+
+<script setup lang="ts">
+</script>
+
+<template>
+      <button type="button" class="btn btn-sm btn-outline-danger" data-action="delete">
+      <i class="fa fa-solid fa-trash" data-action="delete"/>
+    </button>
+</template>